backup_cotizador_v7_20251111_060937/pyproject.toml
[project]
name = "smp"
description = "Sistema de Cotizacion Inteligente"
readme = "README.md"
authors = [{ name = "David Salguedo", email = "david@sayainvestments.co" },{ name = "Guillermo Rebaza", email = "grebaza@gmail.com" }]
dynamic = ["version"]
urls = { Homepage = "https://github.com/gosentio/smp.git" }
keywords = ["cost model", "pricing"]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "psycopg2-binary>=2.9.9",
    "pyodbc>=5.0.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "openpyxl>=3.1.2",
    "httpx>=0.25.2",
    "colorlog>=6.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.0.1",
    "cachetools>=5.3.2",
    "platformdirs>=4.4.0",
]

[project.optional-dependencies]
opt = []

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = { attr = "smp.__version__" }

[dependency-groups]
dev = ["pre-commit>=4.2.0", "pytest>=8.3.5", "pytest-benchmark>=5.1.0"]

[tool.pytest.ini_options]
addopts = "--benchmark-columns=min,mean,stddev,median,max,rounds"
log_cli = true
log_cli_level = "DEBUG"
